AI/requirements.txt
fastapi
tapi
uvicorn
openai
clarifai
nbformat
nbclient
nbconvert
kafka-python